Move things around 8 / 12
Move Resolver to a separate file under ammonite.frontend
